requirements.txt
apache-beam[gcp]==2.56.0
pyarrow==13.0.0
google-cloud-storage==2.16.0